refactor(users): 将tinyint字段改为smallint以支持更大范围
修改用户模块中的字段类型，将status和gender字段从tinyint改为smallint，以支持更大的数值范围。同时，更新了用户列表查询功能，支持分页和筛选条件

+++ sports-hub-project/src/modules/users/entities/role.entity.ts
@@ -33,7 +33,7 @@ export class Role {
   description: string;
 
   @ApiProperty({ description: '角色状态(0:禁用,1:启用)', example: 1 })
-  @Column({ type: 'tinyint', default: 1 })
+  @Column({ type: 'smallint', default: 1 })
   status: number;
 
   @ApiProperty({ description: '创建时间' })
